Finish Part 2

+++ src/day_8/mod.rs
@@ -17,7 +17,7 @@ fn part_1(map: &Map) -> u32 {
     map.steps_between("AAA", "ZZZ")
 }
 
-fn part_2(map: &Map) -> u32 {
+fn part_2(map: &Map) -> usize {
     map.steps_between_all('A', 'Z')
 }
